pyproject.toml
[tool.poetry]
name = "onboarding-test"
version = "0.1.0"
description = ""
authors = ["Aida Paul <aida@reddy.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
black = "^24.3.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
